Move reduce_corpus function to RepoMapping
